+++ test/tx-data-send.test.js
@@ -1,6 +1,6 @@
 import {Buffer} from 'safe-buffer';
 import {mPrefixToHex, convertToPip} from 'minterjs-util';
-import {MinterTx, MinterTxSignature, MinterTxDataSend, TX_TYPE_SEND, formatCoin} from '~/src';
+import {MinterTx, MinterTxSignature, MinterTxDataSend, TX_TYPE_SEND, coinToBuffer} from '~/src';
 
 // exercise fantasy smooth enough arrive steak demise donkey true employ jealous decide blossom bind someone
 const PRIVATE_KEY = new Buffer('5fa3a8b186f6cc2d748ee2d8c0eb7a905a7b73de0f2c34c5e7857c3b46f187da', 'hex');
@@ -15,7 +15,7 @@ const FORM_DATA = {
 describe('tx send', () => {
     const txData = (new MinterTxDataSend({
         to: mPrefixToHex(FORM_DATA.address),
-        coin: formatCoin(FORM_DATA.coin),
+        coin: coinToBuffer(FORM_DATA.coin),
         value: `0x${convertToPip(FORM_DATA.amount, 'hex')}`,
     })).serialize();
 
@@ -25,7 +25,7 @@ describe('tx send', () => {
             nonce: '0x01',
             chainId: '0x01',
             gasPrice: '0x01',
-            gasCoin: formatCoin('MNT'),
+            gasCoin: coinToBuffer('MNT'),
             type: TX_TYPE_SEND,
             data: txData,
             payload: `0x${Buffer.from(FORM_DATA.payload, 'utf-8').toString('hex')}`,
@@ -37,18 +37,18 @@ describe('tx send', () => {
     });
 
     test('tx2 signature', () => {
-        const VALID_SIGNATURE = 'f8840801018a4d4e540000000000000001aae98a4d4e540000000000000094376615b9a3187747dc7c32e51723515ee62e37dc880de0b6b3a7640000808080b845f8431ba00744d2e5dc8abf6b1f5b5ce3540f3552721ced15308da1459fe7b279f3146603a035d9d02a22e5d3f2e591878a3b5350aa6568ade848bde2ccf9a77f81c5ef38fe';
+        const VALID_SIGNED_TX = 'f8840801018a4d4e540000000000000001aae98a4d4e540000000000000094376615b9a3187747dc7c32e51723515ee62e37dc880de0b6b3a7640000808080b845f8431ba00744d2e5dc8abf6b1f5b5ce3540f3552721ced15308da1459fe7b279f3146603a035d9d02a22e5d3f2e591878a3b5350aa6568ade848bde2ccf9a77f81c5ef38fe';
         const txParams = {
             nonce: '0x08',
             chainId: '0x01',
             gasPrice: '0x01',
-            gasCoin: formatCoin('MNT'),
+            gasCoin: coinToBuffer('MNT'),
             type: TX_TYPE_SEND,
             data: txData,
         };
         const tx = new MinterTx(txParams);
         tx.signatureData = (new MinterTxSignature()).sign(tx.hash(false), PRIVATE_KEY).serialize();
-        expect(tx.serialize().toString('hex')).toEqual(VALID_SIGNATURE);
+        expect(tx.serialize().toString('hex')).toEqual(VALID_SIGNED_TX);
     });
 });
 
@@ -63,7 +63,7 @@ describe('tx send (php test)', () => {
     // value: '0x5f5e100'
     const txData = {
         to: mPrefixToHex(TO_ADDRESS),
-        coin: formatCoin('MNT'),
+        coin: coinToBuffer('MNT'),
         value: `0x${convertToPip(1, 'hex')}`,
     };
     // nonce: '0x01',
@@ -73,7 +73,7 @@ describe('tx send (php test)', () => {
     const txParams = {
         nonce: '0x01',
         gasPrice: '0x01',
-        gasCoin: formatCoin('MNT'),
+        gasCoin: coinToBuffer('MNT'),
         type: TX_TYPE_SEND,
         data: (new MinterTxDataSend(txData)).serialize(),
         signatureType: '0x01',
@@ -83,7 +83,7 @@ describe('tx send (php test)', () => {
     tx.signatureData = (new MinterTxSignature()).sign(tx.hash(false), PHP_PRIVATE_KEY).serialize();
 
     test('tx signature', () => {
-        const VALID_SIGNATURE = 'f8840180018a4d4e540000000000000001aae98a4d4e540000000000000094fe60014a6e9ac91618f5d1cab3fd58cded61ee99880de0b6b3a7640000808001b845f8431ca0d4e4bca60bce8ebc50f0a13720550b59c4fcd891044dd688576ee56cd409a8a1a016e2c89754e9c01c7b1390797d56a812e76ef6a0e0ed64d1334c1d48f645f4fd';
-        expect(tx.serialize().toString('hex')).toEqual(VALID_SIGNATURE);
+        const VALID_SIGNED_TX = 'f8840180018a4d4e540000000000000001aae98a4d4e540000000000000094fe60014a6e9ac91618f5d1cab3fd58cded61ee99880de0b6b3a7640000808001b845f8431ca0d4e4bca60bce8ebc50f0a13720550b59c4fcd891044dd688576ee56cd409a8a1a016e2c89754e9c01c7b1390797d56a812e76ef6a0e0ed64d1334c1d48f645f4fd';
+        expect(tx.serialize().toString('hex')).toEqual(VALID_SIGNED_TX);
     });
 });